alpha-focus_on_depth/requirements.txt
torch
torchvision
timm
numpy
pillow
transformers
einops
alphaui
jinja2